Preserve the full TotalTimeSeconds precision in TCX
Qt's default `QString::arg` behaviour is to truncate at 6 signifant
digits, but we often have more available from the V800 protobuf data.

+++ src/polar/v2/trainingsession.h
@@ -152,7 +152,7 @@ private:
 
     void addLapStats(QDomDocument &doc, QDomElement &lap,
                      const QVariantMap &base, const QVariantMap &stats,
-                     const double duration = 0, const double distance = 0) const;
+                     const quint64 duration = 0, const double distance = 0) const;
 
 };
